refactor: 경험 보드 액션 로직 분리
